requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
Click==7.0
cryptography==2.4.2
dnspython==1.16.0
elasticsearch==6.3.1
enum34==1.1.6
Flask==1.0.2
Flask-Cors==3.0.7
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.8
ipaddress==1.0.22
IPy==0.83
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
MySQL-python==1.2.5
mysqlclient==1.3.14
pycparser==2.19
pygeoip==0.3.2
pyOpenSSL==18.0.0
requests==2.12.1
six==1.12.0
SQLAlchemy==1.2.16
urllib3==1.24.1
virtualenv==16.0.0
Werkzeug==0.14.1
redis